C++ conversion

+++ include/libc/stddef.h
@@ -1,9 +1,9 @@
 #ifndef _LIBC_STDDEF_H_
 #define _LIBC_STDDEF_H_
 
-typedef unsigned int size_t;
-typedef int ssize_t;
+typedef unsigned long size_t;
+typedef long ssize_t;
 
-typedef int ptrdiff_t; 
+typedef long ptrdiff_t; 
 
 #endif // _LIBC_STDDEF_H_
